requirements.txt
certifi==2018.10.15
chardet==3.0.4
defusedxml==0.5.0
Django==2.1.3
django-crispy-forms==1.7.2
django-materialize==1.0.1a2
django-mysql==2.4.1
django-paypal==0.5.0
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
idna==2.7
mysql-connector-python==8.0.13
oauthlib==2.1.0
Pillow==5.3.0
protobuf==3.6.1
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.7
requests==2.20.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
urllib3==1.24.1
